+++ src/BiliLite.UWP/Modules/Home/DynamicVM.cs
@@ -495,8 +495,8 @@ namespace BiliLite.Modules
         public string image { get; set; }
         public string image_enhance { get; set; }
         public int expire { get; set; }
-        public int pid { get; set; }
-        public int n_pid { get; set; }
+        public long pid { get; set; }
+        public long n_pid { get; set; }
         public string name { get; set; }
     }
     public class DynamicCardModel2024_decorate
